app/backend/requirements.txt
azure-identity==1.12.0
Flask==2.2.2
langchain==0.0.78
openai==0.26.4
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
flask-restx==1.1.0
